Be explicit about using large groups with -I (#622)
* Be explicit about using large groups with -I

--list-of-ips (-I) only works correctly with groups of order > 2^{32}, since
we have to check each element against a bitmap. This edits the group
selection to always use a large group with using -I.

To make this more obvious to the user, this adds a log_warn if the list
of IPs file is considerably smaller (100000x) than the group size. It
also updates the time remaining calculation to use iterations through
the group when using -I, rather than packets sent.

* Shrink log message to one line

* Update zmap.c

Co-authored-by: Zakir Durumeric <zakird@gmail.com>